+++ src/components/calendar/AddSchedule.jsx
@@ -1,33 +1,29 @@
 import React, { useState } from "react";
 import styled from "styled-components";
 
+import { useNavigate } from "react-router-dom";
 import { useDispatch } from "react-redux";
-import { scheduleUpdate } from "../redux/modules/schedule";
+import { schedulePost } from "../../redux/modules/schedule";
 
 // 스티커
-import img1 from "../assets/img/sticker/sticker1.png";
-import img2 from "../assets/img/sticker/sticker2.png";
-import img3 from "../assets/img/sticker/sticker3.png";
-import img4 from "../assets/img/sticker/sticker4.png";
-import img5 from "../assets/img/sticker/sticker5.png";
-import img6 from "../assets/img/sticker/sticker6.png";
-import img7 from "../assets/img/sticker/sticker7.png";
-import img8 from "../assets/img/sticker/sticker8.png";
-import img9 from "../assets/img/sticker/sticker9.png";
+import img1 from "../../assets/img/sticker/sticker1.png";
+import img2 from "../../assets/img/sticker/sticker2.png";
+import img3 from "../../assets/img/sticker/sticker3.png";
+import img4 from "../../assets/img/sticker/sticker4.png";
+import img5 from "../../assets/img/sticker/sticker5.png";
+import img6 from "../../assets/img/sticker/sticker6.png";
+import img7 from "../../assets/img/sticker/sticker7.png";
+import img8 from "../../assets/img/sticker/sticker8.png";
 
 // 커버 이미지
-import cover1 from "../assets/img/cover/cover1.png";
-import cover2 from "../assets/img/cover/cover2.png";
-import cover3 from "../assets/img/cover/cover3.png";
-import cover4 from "../assets/img/cover/cover4.png";
-import cover5 from "../assets/img/cover/cover5.png";
+import cover1 from "../../assets/img/cover/cover1.png";
 
 // 아이콘
-import time from "../assets/img/icon/Time.svg";
-import location from "../assets/img/icon/Location.svg";
-import memoimg from "../assets/img/icon/memo.svg";
-import emptyImg from "../assets/img/illust/needlogin.png";
-import arrow from "../assets/img/icon/Back.svg";
+import time from "../../assets/img/icon/Time.svg";
+import location from "../../assets/img/icon/Location.svg";
+import memoimg from "../../assets/img/icon/memo.svg";
+import emptyImg from "../../assets/img/illust/needlogin.png";
+import arrow from "../../assets/img/icon/Back.svg";
 
 //Date Picker
 import DatePicker from "react-datepicker";
@@ -35,16 +31,9 @@ import "date-fns";
 import ko from "date-fns/locale/ko";
 import "react-datepicker/dist/react-datepicker.css";
 
-const UpdateSchedule = ({
-  scheduleId,
-  updateScheduleShow,
-  setUpdateScheduleShow,
-  detailInfo,
-  value,
-  onChange,
-  ...others
-}) => {
+const AddSchedule = ({ value, onChange, ...others }) => {
   const dispatch = useDispatch();
+  const navigate = useNavigate();
   const [empty, setEmpty] = useState(true);
 
   //리스트 토글
@@ -54,72 +43,23 @@ const UpdateSchedule = ({
   const [dateShow, setDateShow] = useState(true);
   const [timeShow, setTimeShow] = useState(false);
 
-  function coverimage(detailInfo) {
-    if (detailInfo === 1) {
-      return cover1;
-    } else if (detailInfo === 2) {
-      return cover2;
-    } else if (detailInfo === 3) {
-      return cover3;
-    } else if (detailInfo === 4) {
-      return cover4;
-    } else if (detailInfo === 5) {
-      return cover5;
-    }
-  }
-
-  function colorPickOn(detailInfo) {
-    if (detailInfo === 1) {
-      return "var(--blue1)";
-    } else if (detailInfo === 2) {
-      return "var(--point3)";
-    } else if (detailInfo === 3) {
-      return "rgba(253,187,110,1)";
-    } else if (detailInfo === 4) {
-      return "rgba(253,247,110,1)";
-    } else if (detailInfo === 5) {
-      return "rgba(110,253,150,1)";
-    } else if (detailInfo === 6) {
-      return "rgba(110,218,253,1)";
-    } else if (detailInfo === 7) {
-      return "rgba(130,110,253,1)";
-    } else if (detailInfo === 8) {
-      return "var(--gray2)";
-    } else if (detailInfo === 9) {
-      return "var(--blue4)";
-    }
-  }
-  // console.log("이미지", detailInfo.coverImage);
   //작성목록
-  const [color, setColor] = useState(detailInfo?.color);
-  const [sticker, setSticker] = useState(String(detailInfo?.sticker));
-  const [image, setImage] = useState(detailInfo?.coverImage);
-  const [companyName, setCompanyName] = useState(detailInfo?.companyName);
-  const [title, setTitle] = useState(detailInfo?.title);
-  const [place, setPlace] = useState(detailInfo?.place);
-  const [memo, setMemo] = useState(detailInfo?.memo);
-
-  const [cover, setCover] = useState(coverimage(detailInfo?.coverImage));
-
+  const [color, setColor] = useState("1");
+  const [sticker, setSticker] = useState("1");
+  const [image, setImage] = useState("1");
+  const [companyName, setCompanyName] = useState("");
+  const [title, setTitle] = useState("");
+  const [place, setPlace] = useState("");
+  const [memo, setMemo] = useState("");
+  const [cover, setCover] = useState(cover1);
   //컬러 미리보기
-  const [colorPick, setColorPick] = useState(colorPickOn(detailInfo?.color));
+  const [colorPick, setColorPick] = useState("");
+  const [startDate, setStartDate] = useState(new Date());
 
-  let now = detailInfo?.date;
-  now = now.replace(/-/g, "/");
-  const [startDate, setStartDate] = useState(new Date(now));
-  const hh = detailInfo?.date.split(" ")[1].substr(3, 2);
-  const ttt = detailInfo?.date.split(" ")[1].substr(0, 2);
-  const tt = () => {
-    if (tt > 12) {
-      return "오후";
-    } else {
-      return "오전";
-    }
-  };
   //Time Picker
-  const [selectTime, setSelectTime] = useState(tt(ttt));
-  const [selectHour, setSelectHour] = useState(ttt);
-  const [selectMinute, setSelectMinute] = useState(hh);
+  const [selectTime, setSelectTime] = useState("오전");
+  const [selectHour, setSelectHour] = useState("01");
+  const [selectMinute, setSelectMinute] = useState("00");
 
   //Date Picker
   let [week, month, day, year, sTime] = startDate.toString().split(" ");
@@ -174,7 +114,7 @@ const UpdateSchedule = ({
 
   // 뒤로가기
   const moveBtn = () => {
-    setUpdateScheduleShow(!updateScheduleShow);
+    navigate("/main");
   };
   // 컬러 리스트 보이기
   const colorShowBtn = () => {
@@ -183,13 +123,13 @@ const UpdateSchedule = ({
   // 스티커 리스트 보이기
   const stickerShowBtn = () => {
     setStickerShow(!stickerShow);
-    // setCoverShow(false);
+    setCoverShow(false);
   };
 
   // 커버 이미지 보이기
   const coverShowBtn = () => {
     setCoverShow(!coverShow);
-    // setStickerShow(false);
+    setStickerShow(false);
   };
 
   // 컬러 미리보기 / 컬러 pcik
@@ -204,6 +144,7 @@ const UpdateSchedule = ({
     setSticker(e.target.id);
     // setStickerShow(!stickerShow);
   };
+
   // 커버 이미지 미리보기 / 커버 이미지 pick
   function stickerChangeView(sticker) {
     if (sticker === "1") {
@@ -234,16 +175,23 @@ const UpdateSchedule = ({
   // 커버 이미지 미리보기 / 커버 이미지 pick
   const coverChange = (e) => {
     setCover(e.target.value);
-    setImage(Number(e.target.id));
+    setImage(e.target.id);
     // setCoverShow(!coverShow);
   };
+
   // 일정등록
   const addScheduleBtn = async () => {
-    if ((companyName === "", title === "", allDate === "", place === "")) {
+    if (
+      (companyName === "",
+      title === "",
+      allDate === "",
+      place === "",
+      memo === "")
+    ) {
       setEmpty(!empty);
     } else {
-      await dispatch(
-        scheduleUpdate({
+      dispatch(
+        schedulePost({
           image: Number(image),
           companyName, //필수입력
           title, //필수입력
@@ -252,10 +200,9 @@ const UpdateSchedule = ({
           place, //필수입력
           memo,
           color: Number(color),
-          scheduleId,
         })
       );
-      setUpdateScheduleShow(!updateScheduleShow);
+      navigate("/main");
     }
   };
   const formatDate = (d) => {
@@ -270,10 +217,12 @@ const UpdateSchedule = ({
   };
   const timeShowBtn = () => {
     setDateShow(false);
+
     setTimeShow(!timeShow);
   };
+
   return (
-    <UpdateSchesuleWrap>
+    <AddSchesuleWrap>
       {empty ? (
         ""
       ) : (
@@ -292,7 +241,7 @@ const UpdateSchedule = ({
           </NeedPostModal>
         </NeedPost>
       )}
-      <Header>
+      <Header style={{ backgroundImage: `url(${cover})` }}>
         <AddFlex>
           <Btn onClick={moveBtn}>
             <img src={arrow} alt="뒤로가기" />
@@ -440,7 +389,6 @@ const UpdateSchedule = ({
             )}
           </TitleInput>
         </BtnFlex>
-        <Cover src={cover} />
       </Header>
       <AddList>
         <TitleInput>
@@ -450,10 +398,10 @@ const UpdateSchedule = ({
             onChange={(event) => {
               setCompanyName(event.target.value);
             }}
-            value={companyName}
           />
           <StickerView src={stickerChangeView(sticker)} alt="스티커미리보기" />
         </TitleInput>
+
         <TitleInput>
           <InputText
             type="text"
@@ -461,7 +409,6 @@ const UpdateSchedule = ({
             onChange={(event) => {
               setTitle(event.target.value);
             }}
-            value={title}
           />
           <ColorPicker
             onClick={colorShowBtn}
@@ -471,16 +418,16 @@ const UpdateSchedule = ({
           {colorPickerShow ? (
             <ColorList>
               <Color1 htmlFor="1">
-                <Input
-                  type="radio"
-                  name="color"
-                  id="1"
-                  value="var(--blue1)"
-                  onChange={colorChange}
-                />
                 <ColorCoverLine colorPick={colorPick}></ColorCoverLine>
               </Color1>
-              <Color2 htmlFor="2"></Color2>
+              <Input
+                type="radio"
+                name="color"
+                id="1"
+                value="var(--blue1)"
+                onChange={colorChange}
+              />
+              <Color htmlFor="2" bgColor="var(--point3)"></Color>
               <Input
                 type="radio"
                 name="color"
@@ -488,7 +435,7 @@ const UpdateSchedule = ({
                 value="var(--point3)"
                 onChange={colorChange}
               />
-              <Color3 htmlFor="3"></Color3>
+              <Color htmlFor="3" bgColor="rgba(253, 187, 110, 1)"></Color>
               <Input
                 type="radio"
                 name="color"
@@ -496,7 +443,7 @@ const UpdateSchedule = ({
                 value="rgba(253, 187, 110, 1)"
                 onChange={colorChange}
               />
-              <Color4 htmlFor="4"></Color4>
+              <Color htmlFor="4" bgColor="rgba(253, 247, 110, 1)"></Color>
               <Input
                 type="radio"
                 name="color"
@@ -504,7 +451,7 @@ const UpdateSchedule = ({
                 value="rgba(253, 247, 110, 1)"
                 onChange={colorChange}
               />
-              <Color5 htmlFor="5"></Color5>
+              <Color htmlFor="5" bgColor="rgba(110, 253, 150, 1)"></Color>
               <Input
                 type="radio"
                 name="color"
@@ -512,7 +459,7 @@ const UpdateSchedule = ({
                 value="rgba(110,253,150,1)"
                 onChange={colorChange}
               />
-              <Color6 htmlFor="6"></Color6>
+              <Color htmlFor="6" bgColor="rgba(110, 218, 253, 1)"></Color>
               <Input
                 type="radio"
                 name="color"
@@ -520,7 +467,7 @@ const UpdateSchedule = ({
                 value="rgba(110,218,253,1)"
                 onChange={colorChange}
               />
-              <Color7 htmlFor="7"></Color7>
+              <Color htmlFor="7" bgColor="rgba(130, 110, 253, 1)"></Color>
               <Input
                 type="radio"
                 name="color"
@@ -528,7 +475,7 @@ const UpdateSchedule = ({
                 value="rgba(130,110,253,1)"
                 onChange={colorChange}
               />
-              <Color8 htmlFor="8"></Color8>
+              <Color htmlFor="8" bgColor="var(--gray2)"></Color>
               <Input
                 type="radio"
                 name="color"
@@ -591,7 +538,7 @@ const UpdateSchedule = ({
             </DateWrap>
           )}
           {timeShow && (
-            <Modal2 className="modal">
+            <Modal className="modal">
               <div className="section">
                 <div className="select-time">
                   <div className="division">
@@ -647,7 +594,7 @@ const UpdateSchedule = ({
                   </div>
                 </div>
               </div>
-            </Modal2>
+            </Modal>
           )}
           <PlaceText>
             <input
@@ -656,7 +603,6 @@ const UpdateSchedule = ({
               onChange={(event) => {
                 setPlace(event.target.value);
               }}
-              value={place}
             />
             <img src={location} alt="장소" />
           </PlaceText>
@@ -667,16 +613,15 @@ const UpdateSchedule = ({
               onChange={(event) => {
                 setMemo(event.target.value);
               }}
-              value={memo}
             />
           </TextArea>
         </Pick>
       </AddList>
-    </UpdateSchesuleWrap>
+    </AddSchesuleWrap>
   );
 };
 
-export default UpdateSchedule;
+export default AddSchedule;
 const NeedPost = styled.div`
   position: absolute;
   top: 50%;
@@ -685,7 +630,7 @@ const NeedPost = styled.div`
   width: 100%;
   height: 100vh;
   background-color: rgba(17, 17, 17, 0.3);
-  z-index: 999;
+  z-index: 99999;
 `;
 const NeedPostBtn = styled.button`
   background-color: var(--blue4);
@@ -726,16 +671,12 @@ const NeedPostModal = styled.div`
   width: 45%;
   text-align: center;
 `;
-const UpdateSchesuleWrap = styled.div`
+
+const AddSchesuleWrap = styled.div`
   background-color: var(--blue1);
   width: 100%;
-  height: 100%;
+  height: 100vh;
   font-weight: 500;
-  position: absolute;
-  top: 0;
-  left: 50%;
-  transform: translateX(-50%);
-  z-index: 999;
   input {
     outline: none;
     padding: 18px;
@@ -744,6 +685,7 @@ const UpdateSchesuleWrap = styled.div`
     border: 0;
     outline: none;
     font-weight: 500;
+
     :focus {
       outline: none;
     }
@@ -759,8 +701,9 @@ const Btn = styled.button`
   color: #fff;
   background-color: transparent;
   border: 1px solid #fff;
+  padding: 6px 10px;
   border-radius: 8px;
-  z-index: 999;
+  z-index: 99;
 `;
 const StickerAddBtn = styled.button`
   font-weight: 700;
@@ -818,19 +761,20 @@ const Header = styled.div`
   width: 90%;
   padding: 20px;
   height: 184px;
-  background-size: contain;
+  background-size: cover;
+  background-position: bottom;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
   align-items: center;
-  position: relative;
+  background-repeat: no-repeat;
 `;
 const AddFlex = styled.div`
   width: 100%;
   display: flex;
   justify-content: space-between;
   align-items: center;
-  z-index: 999;
+  z-index: 99;
 `;
 const BtnFlex = styled.div`
   width: 100%;
@@ -845,7 +789,6 @@ const InputText = styled.input`
   padding: 10px;
 `;
 const PlaceText = styled.div`
-  margin-top: 16px;
   position: relative;
   img {
     position: absolute;
@@ -874,18 +817,7 @@ const PlaceText = styled.div`
     width: calc(90% - 24px);
     padding-left: 44px !important;
   }
-  background: url(../assets/img/icon/Location.png) center center no-repeat !important;
-`;
-
-const Background = styled.div`
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  background-color: rgba(0, 0, 0, 0.5);
-  width: 100%;
-  height: 100vh;
-  z-index: 99;
+  background: url(../../assets/img/icon/Location.svg) center center no-repeat !important;
 `;
 const TextArea = styled.div`
   position: relative;
@@ -946,7 +878,16 @@ const ColorPicker = styled.button`
     border: ${(props) => (props.colorPick ? "" : "1px solid var(--blue4)")};
   }
 `;
-
+const Background = styled.div`
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  background-color: rgba(0, 0, 0, 0.5);
+  width: 100%;
+  height: 100vh;
+  z-index: 99;
+`;
 const StickerList = styled.div`
   position: absolute;
   top: 218px;
@@ -978,14 +919,7 @@ const CoverList = styled.div`
   width: 275px;
   z-index: 99;
 `;
-const Cover = styled.img`
-  width: 100%;
-  position: absolute;
-  bottom: 0;
-  left: 50%;
-  transform: translateX(-50%);
-  z-index: 1;
-`;
+
 const ColorList = styled.div`
   display: flex;
   justify-content: space-between;
@@ -1002,94 +936,32 @@ const ColorList = styled.div`
 const Input = styled.input`
   display: none;
 `;
+const Color = styled.label`
+  width: 28px;
+  height: 28px;
+  border-radius: 100%;
+  display: block;
+  border: 2px solid var(--gray2);
+  cursor: pointer;
+  overflow: hidden;
+  background-color: ${(props) => props.bgColor};
+`;
 const Color1 = styled.label`
-  background-color: #fff;
-  width: 30px;
-  height: 30px;
+  width: 28px;
+  height: 28px;
   border-radius: 100%;
   display: block;
   border: 2px solid var(--gray2);
   cursor: pointer;
   overflow: hidden;
 `;
-const Color2 = styled.label`
-  background-color: var(--point3);
-  width: 30px;
-  height: 30px;
-  border-radius: 100%;
-  display: block;
-  border: 2px solid var(--gray2);
-  cursor: pointer;
-`;
-const Color3 = styled.label`
-  background-color: rgba(253, 187, 110, 1);
-  width: 30px;
-  height: 30px;
-  border-radius: 100%;
-  display: block;
-  border: 2px solid var(--gray2);
-  cursor: pointer;
-`;
-const Color4 = styled.label`
-  background-color: rgba(253, 247, 110, 1);
-  width: 30px;
-  height: 30px;
-  border-radius: 100%;
-  display: block;
-  border: 2px solid var(--gray2);
-  cursor: pointer;
-`;
-const Color5 = styled.label`
-  background-color: rgba(110, 253, 150, 1);
-  width: 30px;
-  height: 30px;
-  border-radius: 100%;
-  display: block;
-  border: 2px solid var(--gray2);
-  cursor: pointer;
-`;
-const Color6 = styled.label`
-  background-color: rgba(110, 218, 253, 1);
-  width: 30px;
-  height: 30px;
-  border-radius: 100%;
-  display: block;
-  border: 2px solid var(--gray2);
-  cursor: pointer;
-`;
-const Color7 = styled.label`
-  background-color: rgba(130, 110, 253, 1);
-  width: 30px;
-  height: 30px;
-  border-radius: 100%;
-  display: block;
-  border: 2px solid var(--gray2);
-  cursor: pointer;
-`;
-const Color8 = styled.label`
-  background-color: var(--gray2);
-  width: 30px;
-  height: 30px;
-  border-radius: 100%;
-  display: block;
-  border: 2px solid var(--gray2);
-  cursor: pointer;
-`;
+
 const StickerImg = styled.img`
   width: 50px;
   height: 50px;
   cursor: pointer;
 `;
-const CoverImg = styled.div`
-  width: 50px;
-  height: 50px;
-  overflow: hidden;
-  border-radius: 10px;
-  cursor: pointer;
-  img {
-    height: 100%;
-  }
-`;
+
 const StickerPick1 = styled.label`
   width: 50px;
   height: 50px;
@@ -1114,9 +986,12 @@ const CoverPick1 = styled.label`
   cursor: pointer;
   :nth-child(1) {
     border: ${(props) =>
-      props.image === 1 ? "2px solid var(--blue4)" : "2px solid var(--gray2)"};
-    color: ${(props) => (props.image === 1 ? "var(--blue4)" : "var(--gray2)")};
-    font-weight: ${(props) => (props.image === 1 ? "700px" : "500px")};
+      props.image === "1"
+        ? "2px solid var(--blue4)"
+        : "2px solid var(--gray2)"};
+    color: ${(props) =>
+      props.image === "1" ? "var(--blue4)" : "var(--gray2)"};
+    font-weight: ${(props) => (props.image === "1" ? "700px" : "500px")};
   }
 `;
 const StickerCoverLine = styled.div`
@@ -1139,7 +1014,7 @@ const ColorCoverLine = styled.div`
 `;
 const CoverLine = styled.div`
   border-bottom: ${(props) =>
-    props.image === 1 ? "2px solid var(--blue4)" : "2px solid var(--gray2)"};
+    props.image === "1" ? "2px solid var(--blue4)" : "2px solid var(--gray2)"};
   width: 150%;
   transform: rotate(45deg);
   transform-origin: top left;
@@ -1228,20 +1103,29 @@ const CoverPick = styled.label`
   cursor: pointer;
   :nth-child(2) {
     border: ${(props) =>
-      props.image === 2 ? "2px solid var(--blue4)" : "2px solid var(--gray2)"};
-    color: ${(props) => (props.image === 2 ? "var(--blue4)" : "var(--gray2)")};
+      props.image === "2"
+        ? "2px solid var(--blue4)"
+        : "2px solid var(--gray2)"};
+    color: ${(props) =>
+      props.image === "2" ? "var(--blue4)" : "var(--gray2)"};
     font-weight: ${(props) => (props.image === "2" ? "700px" : "500px")};
   }
   :nth-child(3) {
     border: ${(props) =>
-      props.image === 3 ? "2px solid var(--blue4)" : "2px solid var(--gray2)"};
-    color: ${(props) => (props.image === 3 ? "var(--blue4)" : "var(--gray2)")};
+      props.image === "3"
+        ? "2px solid var(--blue4)"
+        : "2px solid var(--gray2)"};
+    color: ${(props) =>
+      props.image === "3" ? "var(--blue4)" : "var(--gray2)"};
     font-weight: ${(props) => (props.image === "3" ? "700px" : "500px")};
   }
   :nth-child(4) {
     border: ${(props) =>
-      props.image === 4 ? "2px solid var(--blue4)" : "2px solid var(--gray2)"};
-    color: ${(props) => (props.image === 4 ? "var(--blue4)" : "var(--gray2)")};
+      props.image === "4"
+        ? "2px solid var(--blue4)"
+        : "2px solid var(--gray2)"};
+    color: ${(props) =>
+      props.image === "4" ? "var(--blue4)" : "var(--gray2)"};
     font-weight: ${(props) => (props.image === "4" ? "700px" : "500px")};
   }
 `;
@@ -1280,6 +1164,7 @@ const DateWrap = styled.div`
       inset 0px 8px 14px rgba(255, 255, 255, 0.3);
     border-radius: 6.83801px;
     border: 0;
+    margin-bottom: 16px;
   }
   .react-datepicker__month-container {
     width: 100%;
@@ -1357,12 +1242,10 @@ const DateHead = styled.div`
   display: flex;
   justify-content: space-between;
 `;
-
 const DateBtns = styled.div`
   display: flex;
   gap: 18px;
 `;
-
 const DateBtn = styled.button`
   color: var(--gray2);
   width: 16px;
@@ -1375,12 +1258,10 @@ const DateBtn = styled.button`
   background-color: transparent;
   cursor: pointer;
 `;
-
 const DateMonth = styled.p`
   font-weight: 700;
   font-size: 14px;
 `;
-
 const SelectTimeBtn = styled.button`
   background-color: transparent;
   padding: 10px;
@@ -1388,23 +1269,29 @@ const SelectTimeBtn = styled.button`
   font-size: 20px;
   color: ${(props) => props.color && props.color};
 `;
-const Modal2 = styled.div`
+const Modal = styled.div`
   background-color: var(--blue1);
   box-shadow: 0px 14px 24px -4px rgba(117, 146, 189, 0.32),
     inset 0px 8px 14px rgba(255, 255, 255, 0.3);
   border-radius: 6.83801px;
   border: none;
+
   height: 150px;
   overflow: hidden;
   padding: 18px;
   text-align: center;
+  margin-bottom: 16px;
 
   .select-time {
     display: flex;
     align-items: center;
     justify-content: space-between;
     width: auto;
+
     div {
+      :nth-child(1) {
+        justify-content: center;
+      }
       display: flex;
       flex-direction: column;
       align-items: center;
@@ -1416,9 +1303,6 @@ const Modal2 = styled.div`
       flex: 1;
       text-align: center;
       overflow-y: scroll;
-      :nth-child(1) {
-        justify-content: center;
-      }
     }
   }
 `;
