+++ src/components/calendar/DailyList.jsx
@@ -1,15 +1,15 @@
 import React, { Fragment } from "react";
 import styled from "styled-components";
 import { useSelector } from "react-redux";
-import img1 from "../assets/img/sticker/sticker1.png";
-import img2 from "../assets/img/sticker/sticker2.png";
-import img3 from "../assets/img/sticker/sticker3.png";
-import img4 from "../assets/img/sticker/sticker4.png";
-import img5 from "../assets/img/sticker/sticker5.png";
-import img6 from "../assets/img/sticker/sticker6.png";
-import img7 from "../assets/img/sticker/sticker7.png";
-import img8 from "../assets/img/sticker/sticker8.png";
-import img9 from "../assets/img/sticker/sticker9.png";
+import img1 from "../../assets/img/sticker/sticker1.png";
+import img2 from "../../assets/img/sticker/sticker2.png";
+import img3 from "../../assets/img/sticker/sticker3.png";
+import img4 from "../../assets/img/sticker/sticker4.png";
+import img5 from "../../assets/img/sticker/sticker5.png";
+import img6 from "../../assets/img/sticker/sticker6.png";
+import img7 from "../../assets/img/sticker/sticker7.png";
+import img8 from "../../assets/img/sticker/sticker8.png";
+import img9 from "../../assets/img/sticker/sticker9.png";
 import { Link } from "react-router-dom";
 const DailyList = () => {
   const dailySchedule = useSelector((state) => state.schedule.daily);
